client id is client in 2.5!!!
client (usually optional, required for J! 1.7.3 due to a regression; but highly encouraged regardless) - The client ID of the extension, which can be found by looking inside the #__extensions table. To date, use 0 for "site" and 1 for "administrator". Plugins are automatically installed with a client of 0 (site), but you will need to specify the client in an update or it will default to 1 (administrator). Components are automatically installed with a client of 1, which is currently the default.